Avoid a Gdk crash when sys.argv contains strings with surrogates.
In Python 3, Gdk.init_check() encodes sys.argv without handling surrogates,
causing an UnicodeEncodeError exception while importing Gtk.
The new class SanizeArgv implements a mechanism to avoid that:
- The first Gdk/Gtk import performed by a program should be preceded by
  a SanitizeArgv object creation.
- After Gdk/Gtk import, call this object's method restore().

This has to be done in all modules that may perform the initial Gdk/Gtk import,
and the algorithm is not 100% reliable in the sense that 2 different
consecutive strings with surrogates may map to the same display representation.
In any case, this is just a Gdk bug workaround.